pyproject.toml
[project]
name = "alacritty-os-specific-shell-hook"
authors = [{name = "oryon-dominik", email = "oryon@cyberise.de"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flit >=3.5.1",
    "PyYAML >=6.0",
]
classifiers = ["Private :: Do Not Upload"]
dynamic = ["version", "description"]

[project.scripts]
setalacrittyshell = "hook:main"

[tool.flit.module]
name = "hook"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
